Cleanup the file structure of main
